OEQ-2767 - fix: favourite items is missing after import

See merge request edalex-group/development/oeq/openequella!997

+++ Source/Plugins/Core/com.equella.core/src/com/tle/core/favourites/converter/FavouriteItemsConverter.java
@@ -100,7 +100,7 @@ public class FavouriteItemsConverter extends AbstractMigratableConverter<Object>
 
       Map<Long, Long> itemMap = params.getItems();
       Long newItemId = itemMap.get(bookmark.getItem().getId());
-      bookmark.getItem().setId(newItemId.longValue());
+      bookmark.getItem().setId(newItemId);
 
       bookmarkDao.save(bookmark);
       bookmarkDao.flush();
